+++ App.tsx
@@ -1,13 +1,11 @@
-import { useState } from "react";
-import { StyleSheet, Text, View } from "react-native";
-import SelectedGroup from "./component-first/component";
-import ProgressButton from "./component-second/button";
-import StyledTextInput from "./component-third/input";
 import * as React from 'react';
 import { createNativeStackNavigator } from '@react-navigation/native-stack';
 import { NavigationContainer } from "@react-navigation/native";
-import PageFirst from "./page-first";
+import PageFirst from "./screens/Home";
 import { RootStackParamList } from "./types";
+import First from "./screens/First";
+import Second from "./screens/Second";
+import Third from "./screens/Third";
 const Stack = createNativeStackNavigator<RootStackParamList>();
 
 export default function App() {
@@ -15,9 +13,9 @@ export default function App() {
        <NavigationContainer>
       <Stack.Navigator>
         <Stack.Screen name="Home" component={PageFirst} />
-        <Stack.Screen name="First" component={SelectedGroup} />
-        <Stack.Screen name="Second" component={ProgressButton} />
-        <Stack.Screen name="Third" component={StyledTextInput}/>
+        <Stack.Screen name="First" component={First} />
+        <Stack.Screen name="Second" component={Second} />
+        <Stack.Screen name="Third" component={Third}/>
       </Stack.Navigator>
     </NavigationContainer>
   );
