+++ components/ProgressButton/ProgressButton.tsx
@@ -32,11 +32,6 @@ export default function ProgressButton() {
   });
 
   return (
-    <View style={{ flex: 1,
-      alignItems: "center",
-      justifyContent: "center",
-      backgroundColor: "#5A3ECD",
-      }}>
     <TouchableOpacity style={styles.container} onPress={animateProgress}>
       <View
         style={{
@@ -67,7 +62,7 @@ export default function ProgressButton() {
           },
         ]}
       ></Animated.View>
-    </TouchableOpacity></View>
+    </TouchableOpacity>
   );
 }
