+++ package.json
@@ -12,8 +12,8 @@
   "dependencies": {
     "@react-navigation/native": "*",
     "@react-navigation/native-stack": "*",
-    "react": "18.2.0",
-    "react-native": "0.71.6",
+    "react": "^18.2.0",
+    "react-native": "^0.71.7",
     "react-native-gesture-handler": "^2.9.0",
     "react-native-reanimated": "^3.1.0",
     "react-native-safe-area-context": "*",
