+++ screens/Home.tsx
@@ -1,9 +1,10 @@
 import { Button, StyleSheet, Text, View } from "react-native";
-import { RootStackParamList } from "./types";
+import { RootStackParamList } from "../types";
 import { NativeStackScreenProps } from "@react-navigation/native-stack";
+import React from "react";
 
 type prop = NativeStackScreenProps<RootStackParamList, 'Home'>;
-export default function PageFirst({navigation}:prop) {
+export default function Home({navigation}:prop) {
   return (
     <View style={styles.container}>
        <Button
